move renderable objects to separate files
